Update _tools.py
Weird bug that appear because the junction length reported in the table and the actual junction length is not the same, intereferring with how clones are separated. This is now changed by default to recalculate the junction length within find clones (doesn't update the vdj table as it will be too time consuming to search within this function)

I should do that in the preprocessing perhaps.