Update _diversity.py
Added keey_added and expanded_only options to clone_diversity to be more flexible in some of the sub functions.

Reworked the clone_size gini measure to now use the network as default unless specified otherwise.

Still need to work on a version that runs clone_size gini in expanded only mode.