bump
Deprecate filter_contigs in favour of check_contigs which now has the ability to consider the consensus counts for retaining/filtering contigs.

Update functions and tests that relied on filter_contigs to now just use check_contigs.

Change preprocessing defaults to using 'all' instead of 'filtered' to be consistent with the singularity container.

Deprecate the notebook to run in R since I haven't caught up with how to make reticulate play nice.